+++ src/IEC6205675.h
@@ -21,8 +21,8 @@ private:
     CosemData* getCosemDataAt(uint8_t index, const char* ptr);
     CosemData* findObis(uint8_t* obis, int matchlength, const char* ptr);
     uint8_t getString(uint8_t* obis, int matchlength, const char* ptr, char* target);
-    double getNumber(uint8_t* obis, int matchlength, const char* ptr);
-    double getNumber(CosemData*);
+    float getNumber(uint8_t* obis, int matchlength, const char* ptr);
+    float getNumber(CosemData*);
     time_t getTimestamp(uint8_t* obis, int matchlength, const char* ptr);
 
     uint8_t AMS_OBIS_VERSION[4]                 = {  0, 2, 129, 255 };
